backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pillow>=11.0.0
ollama==0.2.1
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.4
anthropic==0.28.0
openai==1.35.0
supabase==2.5.0
stripe==9.0.0
celery==5.4.0
redis==5.0.0
websockets==12.0
twilio>=9.0.0
pydub>=0.25.1
pytest==8.2.0
pytest-asyncio==0.23.0
